requirements.txt
amqp==5.0.6
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.2
Django==3.2.6
django-autoslug==1.9.8
django-crispy-forms==1.12.0
django-environ==0.8.1
django-phonenumber-field==6.0.0
djangorestframework==3.12.4
django-dotenv==1.4.2
filelock==3.0.12
gunicorn==20.1.0
idna==3.3
kombu==5.1.0
Languages==1.0.0
phonenumbers==8.12.38
platformdirs==2.3.0
prompt-toolkit==3.0.20
django-ckeditor
django-taggit
# psycopg==3.0.8
# psycopg-binary==3.0.8
psycopg2-binary>=2.9
pycparser==2.20
pyttsx3==2.90
pytz==2021.1
rabbitmq==0.2.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.8.0
wcwidth==0.2.5
dj-database-url
whitenoise
uWSGI>=2.0.20,<2.1
